requirements.txt
Django
MySQL-python
South
PIL
django-css
BeautifulSoup
django-cms==2.1.0.beta2
django-reversion
sorl-thumbnail
